feat: add keyword and semantic extractors, and introduce a shared ABC for all extractors
Add extractors to extract attributes (boolean Truth) on the basis of keyword matching (plain string matching) or semantic (matching with embeddings). Create an ABC to deal with common extraction logic, and hook up extraction methods to their extractor with an extractor registry.

- extract shared config and logic out of llm_data_extractor.py into base_extractor.py. LLMDataExtractor now just extends the ABC
- Add KeywordDataExtractor  (substring matching) and SemanticKeywordExtractor (sentence-transformers similarity)
- add extractor_registry.get_data_extractor(config), to map ExtractionMethod and return a config-instantiated instance
- add Keyword annotation type
- add sentence-transformers dependency (lightweight since we already have transformers)
- update tests
