+++ pyproject.toml
@@ -52,6 +52,7 @@ dependencies = [
     "pyyaml>=6.0.2",
     "rapidfuzz>=3.9.0",
     "scikit-learn>=1.7.1",
+    "sentence-transformers>=5.5.1",
     "tabulate>=0.9.0",
     "toml>=0.10.2",
     "typer>=0.21.2",
